Restructure batch decryption to avoid combinatorial explosion.
While it is necessary to perform `m * n` decryptions, where `m` is the
number of outputs being decrypted and `n` is the number of IVKs, it's
inconvenient and unnecessary to return the full cartesian product of
these results, as only one IVK will decrypt a given output. Instead,
when a decryption succeeds, we return the index of the input IVK along
with the output it decrypted.