CI: Replace most unmaintained `actions-rs/*` actions
The actions-rs/* actions are unmaintained. See actions-rs/toolchain#216
for more information.

- actions-rs/toolchain is replaced by rust-toolchain.toml.
- actions-rs/cargo is replaced by direct cargo usage.
- actions-rs/tarpaulin is replaced by the cargo-tarpaulin container.

We currently retain a few usages due to a current lack of alternatives:

- actions-rs/toolchain to enable beta (for Clippy warnings) and nightly
  (for building the book).
- actions-rs/clippy-check so Clippy results are published as annotations
  on PRs.

We also now require via rust-toolchain.toml that people building from
the repo have the clippy and rustfmt components installed. This is fine
because they will need to ensure these pass when making PRs.