Reduce the size of proptest-generated history trees.
After replacing `quickcheck` with `proptest` in `zcash_history`, memory
consumption of the property tests grew substantially as a consequence of
sample generation exploring the upper bounds of sample ranges;
previously, the large operation counts would have occurred only with low
probability. However, the correctness of the operations on the `Tree`
data structure should not depend upon the tree size, so it's fine to
reduce the number of operations in order to bound the memory
consumption. As the documentation on `Tree` notes, this data structure
is not intended to be used stand-alone because doing so may result in
unbounded memory use; the property tests were doing exactly this and so
were producing this exact pathological memory behavior.