Use mermaid for migrations graph
The migrations ASCII art previously used was somewhat hard to read and much harder to maintain over time.
Mermaid markdown provides a very easy way to add new nodes to the graph, and the result is far more visually parseable.

Mermaid can be rendered on GitHub or via a local extension such as in VS Code. I've added an extension recommendation for VS Code as part of this change.