zcash_client_backend: Update default gap limit for external to 10 addresses.
The rationale behind this change is that current implementations of UTXO
retrieval do not attempt to avoid revealing that multiple addresses
belong to the same wallet when using a light wallet server to check for
transparent UTXOs. A comprehensive solution to this problem requires
changes to the light wallet protocol such that wallets no longer need to
submit batches of addresses to the light wallet server, and can instead
determine whether outputs exist for their addresses locally.

By setting the default gap limit to 10 addresses, we make it possible
for wallets that add privacy-preserving UTXO checking in the future
(either by connecting to full node indexes directly in a
privacy-preserving manner or via changes that take advantage of an
updated light wallet protocol) to then start generating new addresses
within a range of the industry-standard 20 address gap limit that have
not yet been revealed to belong to the wallet.