zcash_client_backend: Add `WalletRead::utxo_query_height`
In the presence of gap-limit addresses, querying for UTXOs based
exclusively upon the last height at which such a query was executed is
inadequate to fully recover the contents of a wallet; when a transaction
is discovered in the past, this may advance the gap limit, and newly
exposed addresses may receive outputs immediately following the
discovered transaction. We can expect that when a transparent output
belonging to the wallet is discovered, that the height at which that
transaction was mined was checked in a query that included _all_ of
addresses having child indices less than or equal to that of the
involved address. Therefore, it's safe to start looking for addresses in
the gap as of that point.