Use zcash_script in PCZTs
This replaces the local script implementation with the one from the
zcash_script crate.

The `Script` type that was here is replaced by a family of three types
- `script::PubKey`;
- `script::Sig<PushValue>` – a v5+ script_sig, used for new sigs;
- `script::Sig<Opcode>` – a pre-v5 script\_sig, used for sigs read off
  the chain, which may predate NU5; and
- an unwrapped `Vec<u8>` for script code.

This uses a number of zcash_script features that haven’t yet made it
into the master branch, let alone a release, so this should not be
merged.