zcash_keys: Replace the `UnifiedAddressRequest` struct with an enum.
This change clarifies APIs that take `UnifiedAddressRequest` values by
making it possible to explicitly request that addresses be created using
all available key items. Previously, this was indicated by providing a
`None` value for an `Option<UnifiedAddressRequest>`, but in some
circumstances the meaning of this `None` could potentially be unclear.
Replacing `UnifiedAddressRequest` with an enum that can express the
caller's intent directly simplifies and clarifies a fair amount of code.