zcash_client_sqlite: Expose error details from `get_address_for_index`
This is an API that has several possible failure modes, and we want to
expose some of these to Zallet users (as they were via `zcashd`'s RPC)
that were being merged into an `Ok(None)`. This does mean the return
value of the API is now backend-specific (in that the backend decides
which of two buckets to place failure cases into for backend-agnostic
users), but this API is for more advanced use cases anyway and is more
likely to be called in a backend-specific context, where the error type
is introspectable.

Co-authored-by: Kris Nuttycombe <kris@nutty.land>