zcash_client_backend: Fix bugs that prevented wallet-emptying transactions.
This fixes a bug in the change-splitting logic that prevented the
creation of transactions that could fully empty the wallet. Prior to
this change, the wallet would return an incoorect `InsufficientFunds`
error in the case that the fee that would be required to perform change
splitting could not be paid, even if such splitting would never occur
because there was inadequate change value to meet the splitting
threshold.