zcash_client_backend: Fix overly-constrained type signature of TestState::create_proposed_transaction
The existing type signature of this method made it impossible to use it
with a proposal created by `zcash_client_backend::data_api::wallet::propose_shielding`.

This is technically a semver-breaking change, but as it only affects
test code under the `test-dependencies` feature flag, it's worth it to
include this in a patch release.