zcash_primitives: Make `Transaction::fee_paid` internally resilient to missing inputs.
Previously, the closure passed to `Transaction::fee_paid` that performs
input value retrieval did not provide any way to directly report that a
requested UTXO could not be found, and consequently this absence had to
be handled via the error pathway. This change modifies both the type of
this function, and the result type of `fee_paid` to reflect this common
circumstance. In addition, this change modifies the overall API to
ensure that negative input amounts and negative fees are not
representable.