release script: Make the changelog update logic more robust using markdown parsing.
Prompt:

The next goal is to make the `update_changelog` and
`update_dependent_changelogs` functions more robust: `update_changelog`
needs to operate properly in the case that there does not yet exist a
`PLANNED` section for the release we're about to make. The
`update_dependent_changelogs` function may need to add a `### Changes`
section for the current release if one doesn't exist yet, and may need
to add a `Migrated to` line to that section if no such line exists
recording updated dependencies. Use the `marko` markdown parser and
update the AST directly instead of performing direct manipulation of the
CHANGELOG text, taking advantage of the structured nature of the
CHANGELOG data.