zcash_client_sqlite: Ensure sufficient confirmations before shielding outputs are spendable.
Transparent UTXOs are always considered untrusted, but we allow
zero-conf shielding transactions; in the case that we shield with fewer
than `untruste` confirmations, the resulting shielding outputs must be
treated in the same fashion as the transparent UTXOs that we shielded.

This change causes the maximum receipt height among shielded transparent
UTXOs to be tracked along with each note, and we then use that height
for trust checks.