zcash_client_sqlite: Fix a data consistency issue in v_transactions.expired_unmined
Prior to this change, if a transparent UTXO was detected but the
corresponding block had not been scanned by the wallet, it would be
possible for the `v_transactions` view to incorrectly report the UTXO
had expired without having been mined.

There is a remaining question to be resolved as to why the corresponding
block might have never made it into the blocks table, but this
change should resolve the immediate issue for users observing this
problem.